Quality: New project architecture.
